feat(verification): move verification service from root to auth
