+++ apps/api/src/auth/auth.module.ts
@@ -4,12 +4,12 @@ import { AuthService } from './auth.service';
 import { UserModule } from '../user/user.module';
 import { JwtModule } from '@nestjs/jwt';
 import { MailerModule } from '../mailer/mailer.module';
-import { VerificationModule } from '../verification/verification.module';
 import { JwtStrategy } from './strategies/jwt.stategy';
 import { LocalStrategy } from './strategies/local.strategy';
 import { GithubStrategy } from './strategies/github.strategy';
 import { PassportModule } from '@nestjs/passport';
 import { ConfigModule, ConfigService } from '@nestjs/config';
+import { VerificationModule } from './verification/verification.module';
 
 @Module({
   imports: [
